Move testCase endpoint to Tapir
