+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ManagementResources.scala
@@ -14,7 +14,6 @@ import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.DeploymentUpdateStrategy.StateRestoringStrategy
 import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
 import pl.touk.nussknacker.engine.deployment.LatestVersion
-import pl.touk.nussknacker.engine.test.testcase.TestCase
 import pl.touk.nussknacker.restmodel.{
   CancelRequest,
   DeployRequest,
@@ -32,7 +31,6 @@ import pl.touk.nussknacker.ui.process.deployment._
 import pl.touk.nussknacker.ui.process.deployment.LoggedUserConversions.LoggedUserOps
 import pl.touk.nussknacker.ui.process.processingtype.provider.ProcessingTypeDataProvider
 import pl.touk.nussknacker.ui.process.test.{ScenarioTestService, SerializedScenarioRecordsContent}
-import pl.touk.nussknacker.ui.process.test.testcase.ScenarioWithTestCase
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -219,39 +217,6 @@ class ManagementResources(
                 }
               }
             }
-          }
-        } ~
-        path("testCase" / ProcessNameSegment) { processName =>
-          (post & processDetailsForName(
-            processName
-          ) & skipResultsPerTransitionQueryParam & skipResultsPerNodeQueryParam & entity(as[ScenarioWithTestCase])) {
-            (details, skipResultsPerTransition, skipResultsPerNode, scenarioWithTestCase) =>
-              canDeploy(details.idWithNameUnsafe) {
-                complete {
-                  measureTime("testCase", metricRegistry) {
-                    scenarioTestServices
-                      .forProcessingTypeUnsafe(details.processingType)
-                      .performTestCase(
-                        scenarioWithTestCase.scenario,
-                        details.processVersionUnsafe,
-                        details.isFragment,
-                        scenarioWithTestCase.testCase
-                      )
-                      .flatMap {
-                        case Left(error) =>
-                          Future.failed(PerformTestDesignerError(TestingApiErrorMessages.from(error)))
-                        case Right(value) =>
-                          mapResultsToHttpResponse(
-                            ResultsWithCountsDto.from(
-                              resultsWithCounts = value,
-                              skipResultsPerNode = SkipResultsPerNode(skipResultsPerNode),
-                              skipResultsPerTransition = SkipResultsPerTransition(skipResultsPerTransition),
-                            )
-                          )
-                      }
-                  }
-                }
-              }
           }
         } ~
         // TODO: maybe Write permission is enough here?
