+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ManagementResourcesSpec.scala
@@ -482,237 +482,8 @@ class ManagementResourcesSpec
     }
   }
 
-  test("run test case") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"startProcess","variables":{"input":["ala"]}},
-        |  {"sourceId":"startProcess","variables":{"input":["bela"]}}
-        |]""".stripMargin
-    saveCanonicalProcessAndAssertSuccess(ProcessTestData.sampleScenario)
-
-    val testCase = TestCase(
-      UUID.randomUUID(),
-      "dummy",
-      testDataContent,
-      Map.empty,
-      Map(
-        NodeId("endsuffix") -> List(
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "'ala'".spel,
-            "#contexts[0].input[0]".spel,
-          ),
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "'ala'".spel,
-            "#contexts[1].input[0]".spel,
-          ),
-          // The output variable produced by the messagesuffix node is not visible at that node, but rather at the subsequent one (endsuffix).
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "{message: 'message'}".spel,
-            "#contexts[0].output".spel,
-          ),
-        ),
-        NodeId("messagesuffix") -> List(
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "'ala'".spel,
-            "#contexts[0].input[0]".spel,
-          ),
-        )
-      )
-    )
-    runTestCase(ProcessTestData.sampleScenario, testCase) ~> check {
-      status shouldEqual StatusCodes.OK
-
-      val responseJson = responseAs[Json]
-      responseJson.hcursor
-        .downField("assertionsResults")
-        .downField("endsuffix")
-        .downN(0)
-        .downField("type")
-        .focus shouldBe Some(Json.fromString("SuccessfulAssertion"))
-
-      responseJson.hcursor
-        .downField("assertionsResults")
-        .downField("endsuffix")
-        .downN(1)
-        .focus shouldBe Some(
-        Json.obj(
-          "type"    -> Json.fromString("FailedAssertion"),
-          "message" -> Json.fromString("Expected: ['ala'] but found ['bela']")
-        )
-      )
-
-      responseJson.hcursor
-        .downField("assertionsResults")
-        .downField("endsuffix")
-        .downN(2)
-        .downField("type")
-        .focus shouldBe Some(Json.fromString("SuccessfulAssertion"))
-
-      responseJson.hcursor
-        .downField("assertionsResults")
-        .downField("messagesuffix")
-        .downN(0)
-        .downField("type")
-        .focus shouldBe Some(Json.fromString("SuccessfulAssertion"))
-    }
-  }
-
-  test("run test case with mocked service") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"startProcess","variables":{"input":["ala"]}}
-        |]""".stripMargin
-
-    val scenario = ScenarioBuilder
-      .streaming(ProcessTestData.sampleProcessName.value)
-      .parallelism(1)
-      .source("startProcess", "csv-source")
-      .enricher("someEnricher", "out1", "paramService", "param" -> "'a'".spel)
-      .filter("input", "#input != null".spel)
-      .to(
-        GraphBuilder
-          .buildVariable("message" + "suffix", "output", "message" -> "'message'".spel)
-          .emptySink(
-            "end" + "suffix",
-            "kafka-string",
-            TopicParamName.value     -> "'end.topic'".spel,
-            SinkValueParamName.value -> "#output".spel
-          )
-      )
-
-    saveCanonicalProcessAndAssertSuccess(scenario)
-
-    val testCase = TestCase(
-      UUID.randomUUID(),
-      "dummy",
-      testDataContent,
-      mocks = Map(
-        NodeId("someEnricher") -> EnricherMock("'b'".spel)
-      ),
-      assertions = Map(
-        NodeId("endsuffix") -> List(
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "'b'".spel,
-            "#contexts[0].out1".spel,
-          ),
-        )
-      )
-    )
-    runTestCase(scenario, testCase) ~> check {
-      status shouldEqual StatusCodes.OK
-
-      val responseJson = responseAs[Json]
-      responseJson.hcursor
-        .downField("assertionsResults")
-        .downField("endsuffix")
-        .downN(0)
-        .downField("type")
-        .focus shouldBe Some(Json.fromString("SuccessfulAssertion"))
-    }
-  }
-
-  test("assertion on not existing node should return bad request") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"startProcess","variables":{"input":["ala"]}},
-        |  {"sourceId":"startProcess","variables":{"input":["bela"]}}
-        |]""".stripMargin
-    saveCanonicalProcessAndAssertSuccess(ProcessTestData.sampleScenario)
-
-    val invalidTestCase = TestCase(
-      UUID.randomUUID(),
-      "dummy",
-      testDataContent,
-      mocks = Map.empty,
-      assertions = Map(
-        NodeId("someNotExistingNode") -> List(
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "'ala'".spel,
-            "#contexts[0].input[0]".spel,
-          ),
-        )
-      )
-    )
-    runTestCase(ProcessTestData.sampleScenario, invalidTestCase) ~> check {
-      status shouldEqual StatusCodes.BadRequest
-      responseAs[String] shouldBe "Assertions configured for not existing nodes: someNotExistingNode"
-    }
-  }
-
-  test("mock on not existing node should return bad request") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"startProcess","variables":{"input":["ala"]}},
-        |  {"sourceId":"startProcess","variables":{"input":["bela"]}}
-        |]""".stripMargin
-    saveCanonicalProcessAndAssertSuccess(ProcessTestData.sampleScenario)
-
-    val invalidTestCase = TestCase(
-      UUID.randomUUID(),
-      "dummy",
-      testDataContent,
-      mocks = Map(
-        NodeId("notExistingEnricher") -> EnricherMock("'b'".spel)
-      ),
-      assertions = Map.empty
-    )
-    runTestCase(ProcessTestData.sampleScenario, invalidTestCase) ~> check {
-      status shouldEqual StatusCodes.BadRequest
-      responseAs[String] shouldBe "Mocks configured for not existing nodes: notExistingEnricher"
-    }
-  }
-
-  test("assertion on non existing context variable should return bad request") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"startProcess","variables":{"input":["ala"]}}
-        |]""".stripMargin
-    saveCanonicalProcessAndAssertSuccess(ProcessTestData.sampleScenario)
-
-    val invalidTestCase = TestCase(
-      UUID.randomUUID(),
-      "dummy",
-      testDataContent,
-      Map.empty,
-      Map(
-        NodeId("messagesuffix") -> List(
-          // The output variable produced by the messagesuffix node is not visible at that node, but rather at the subsequent one.
-          PredicateAssertion(
-            Assertion.AssertionOperator.Equals,
-            "{message: 'message'}".spel,
-            "#contexts[0].output".spel,
-          ),
-        )
-      )
-    )
-    runTestCase(ProcessTestData.sampleScenario, invalidTestCase) ~> check {
-      status shouldEqual StatusCodes.BadRequest
-      responseAs[String] should include("There is no property 'output' in type")
-    }
-  }
-
-  test("running test case with invalid scenario should return bad request") {
-    val testDataContent =
-      """[
-        |  {"sourceId":"source","variables":{"input":["ala"]}},
-        |  {"sourceId":"source","variables":{"input":["bela"]}}
-        |]""".stripMargin
-    saveCanonicalProcessAndAssertSuccess(ProcessTestData.invalidProcess)
-
-    val testCase = TestCase(UUID.randomUUID(), "dummy", testDataContent, Map.empty, Map.empty)
-    runTestCase(ProcessTestData.invalidProcess, testCase) ~> check {
-      status shouldEqual StatusCodes.BadRequest
-      val responseAsString = responseAs[String]
-      // todo: should we return something more structured than toString of errors?
-      responseAsString.startsWith("Only scenario without validation errors can be tested. Errors: ") shouldBe true
-    }
-  }
+  // Tests for the testCase endpoint were migrated to ManagementApiHttpServiceBusinessSpec
+  // using the new POST /api/scenarioTesting/{scenarioName}/testCase endpoint.
 
   test("nodeTransitionResults returned in test results should include null variables") {
     val scenario = ScenarioBuilder
